CRUDs necesarios para la cracion de productos con inventarios LISTAR TODOS LOS PRODUCTOS CON INVENTARIOS EN TRUE ACTUALIZAR PRODUCTO CON INVENTARIO EN TRUE ELIMINACION LOGICA CAMBIANDO ACTIVO A TRUE OTROS METODOS

+++ backend/src/main/java/KuHub/modules/inventario/entity/Inventario.java
@@ -19,11 +19,11 @@ public class Inventario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_inventario")
-    private Long idInventario;
+    private Integer idInventario;
 
     @NotNull(message = "El producto es obligatorio")
     @Column(name = "id_producto", nullable = false, unique = true)
-    private Long idProducto;
+    private Integer idProducto;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "id_producto", insertable = false, updatable = false)
